Split DataLoader Library
